Refactored addListItem
- changed ambiguous obj and type input arguments names
- replaced boolean expressions with better descriptors
